refactor(DisplaySwitches) moved to separate component for readability
