+++ src/views/DetailedView/DisplaySwitches/DisplaySwitches.styled.tsx
@@ -1,0 +1,11 @@
+import { Button, Paper, styled } from "@mui/material";
+
+export const DisplaySwitchesContainer = styled(Paper)`
+  display: flex;
+  justify-content: space-around;
+`;
+
+export const DisplaySwitchWrapper = styled(Button)`
+  text-transform: lowercase;
+  line-height: 150%;
+`;
